add tests of dataframe creators
